Final version

+++ src/main/java/servlets/AttendanceAdminServlet.java
@@ -9,6 +9,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.io.IOException;
 
 @WebServlet("/Zavod/admin/attendance")
@@ -25,10 +26,16 @@ public class AttendanceAdminServlet extends HttpServlet {
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        AttendanceDao attendanceDao = new AttendanceDao();
-        attendanceDao.delete(Long.parseLong(req.getParameter("id")));
-        Long idWorker = Long.parseLong(req.getParameter("idworker"));
-        req.setAttribute("attendances", attendanceDao.findByIdWorker(idWorker));
-        req.getRequestDispatcher("attendance.jsp").forward(req, resp);
+        if(!req.getParameter("action").equals("exit")) {
+            AttendanceDao attendanceDao = new AttendanceDao();
+            attendanceDao.delete(Long.parseLong(req.getParameter("id")));
+            Long idWorker = Long.parseLong(req.getParameter("idworker"));
+            req.setAttribute("attendances", attendanceDao.findByIdWorker(idWorker));
+            req.getRequestDispatcher("attendance.jsp").forward(req, resp);
+        } else {
+            HttpSession session = req.getSession(false);
+            session.invalidate();
+            resp.sendRedirect("/Zavod/login");
+        }
     }
 }
